Move all qcodes queries from sqlite_base to sqlite.queries
sqlite_base is left now empty but with all the imports that it used
to have for backwards compatibility.
